add score to players, and display it in the widget. change all playerId variables to qint8 Closes #30

+++ gui/GameArena.h
@@ -109,7 +109,7 @@ public slots:
 
 signals:
     void sigTimeUpdated(int valueInSeconds);
-    void sigNewPlayerGraphic(int,const QPixmap &);
+    void sigNewPlayerGraphic(qint8,const QPixmap &);
 };
